app/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi[all]==0.75.0
uvicorn[standard]==0.15.0
gunicorn==20.1.0
aerich==0.6.3
asyncpg==0.23.0
bcrypt==3.2.0
passlib==1.7.4
tortoise-orm==0.19.0
SQLAlchemy==1.4.32